Split off HueRangePicker in its own file
